Always crop the SVG file
Previously, the SVG width/height would be same as that
of the manim scene. This commit changes that to the
width/height of the object, instead. This should
avoid unnecessary whitespace in the SVG file.

This commit introduces two parameters to `to_svg` method:
- `crop` - It is a boolean value that determines whether
  to crop the SVG file or not.
- `padding` - It is a float value that determines the
  padding to be added to the SVG file.

The previous behavior can be restored by passing
`crop=False` to `to_svg` method.

Fixes https://github.com/naveen521kk/manim-mobject-svg/issues/7

Signed-off-by: Naveen M K <naveen521kk@gmail.com>